+++ packages/doars-navigate/dst/doars-navigate.js
@@ -372,7 +372,7 @@ var navigate_default = (options) => {
       options.cacheInterval
     );
   };
-  const getFromUrl = (url, dispatchEvent) => {
+  const getFromUrl = (url, dispatchEvent, headers) => {
     return new Promise((resolve) => {
       if (window.location.hostname !== url.hostname) {
         resolve(null);
@@ -401,8 +401,13 @@ var navigate_default = (options) => {
       dispatchEvent("-started", {
         url
       });
-      fetch(url, options.defaultInit).then((response) => {
-        if (response.status < 200 || response.status >= 300) {
+      fetch(
+        url,
+        Object.assign({}, options.fetchOptions, {
+          headers: Object.assign({}, headers, options.fetchOptions.headers)
+        })
+      ).then((response) => {
+        if (response.status < 200 || response.status >= 500) {
           dispatchEvent("-failed", {
             response,
             url
@@ -481,6 +486,9 @@ var navigate_default = (options) => {
       if (modifiers.capture) {
         listenerOptions.capture = true;
       }
+      const fetchHeaders = {
+        [libraryOptions.prefix + "-request"]: directive
+      };
       const dispatchEvent = (suffix = "", data = {}) => {
         element.dispatchEvent(
           new CustomEvent(
@@ -505,7 +513,7 @@ var navigate_default = (options) => {
           processExpression,
           transitionIn
         );
-        getFromUrl(url, dispatchEvent).then((result) => {
+        getFromUrl(url, dispatchEvent, fetchHeaders).then((result) => {
           if (!attribute[NAVIGATE].identifier || attribute[NAVIGATE].identifier !== identifier) {
             return;
           }
@@ -633,7 +641,7 @@ var navigate_default = (options) => {
             return;
           }
           const href = anchor.getAttribute("href");
-          getFromUrl(new URL(href, window.location), dispatchEvent);
+          getFromUrl(new URL(href, window.location), dispatchEvent, fetchHeaders);
         };
         element.addEventListener(
           "focusin",
@@ -665,7 +673,8 @@ var navigate_default = (options) => {
                     anchor.target.getAttribute("href"),
                     window.location
                   ),
-                  dispatchEvent
+                  dispatchEvent,
+                  fetchHeaders
                 );
               }
             }