+++ packages/doars-navigate/CHANGELOG.md
@@ -1,5 +1,10 @@
 # CHANGELOG.md
 
+## 2.1.2
+
+- Added `d-request` header to every request denoting the directive name used.
+- Fix issue where fetch options weren't applied.
+
 ## 2.1.1
 
 - Fixed issue with morph modifier.